+++ changelog.md
@@ -2,6 +2,7 @@
 - Fixed missing `onServerShutdown` events in `main` method
 - Fixed missing `onPlayerBan` events when using commands
 - Fixed missing `onPlayerKick` events when using commands
+- Fixed loading of plugins for the co-op server
 
 # FluxLoader v0.6.0
 - Added `onExecute` and `onTerminate` events for plugins. On the server: `onExecute` is called at startup and `onTerminate` - at shutdown. On the client: `onExecute` and `onTerminate` called when entering a single player game (if enabled in the main menu).